lab1/prescription_management/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.3
python-dotenv==1.0.0
requests==2.31.0
